// types.ts
export interface Notification {
  id: string;
  title: string;
  title_bn?: string;
  message: string;
  message_bn?: string;
  timestamp: number;
  type: 'success' | 'failed' | 'admin' | 'bonus' | 'offer';
  targetUid?: string; 
  isAuto?: boolean; 
}